Refactored: Datadog::Event --> Datadog::Core::Event
